Refactor code for controllers
